services/ai-engine/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
pydantic==2.6.0
langchain==0.1.5
langchain-google-genai
pinecone-client==3.0.0
requests==2.31.0
pdfplumber==0.10.3
python-docx==1.1.0
tiktoken==0.5.2
pika==1.3.2
weaviate-client==3.26.0
langchain-community==0.0.17
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
